[refactor] Removes http logic from components. Reworks testing
