+++ src/app/care-form/care-form.component.html
@@ -126,10 +126,11 @@
         </div>
       </div>
     </div>
+  
+    <div *ngIf="changed">
+      <button class="add">Submit Changes</button>
+    </div>
     <div>
-      <div *ngIf="changed">
-        <button class="add">Submit Changes</button>
-      </div>
     </div>
   </div>
 </form>
