+++ src/app/forest/forest.component.html
@@ -26,12 +26,13 @@
                 </li>
 
                 <li class="list-group-item {{ plant.alert.color }}">
-                  <p class="card-text">
-                    Last Watered: {{ plant.daysSinceLastWatered }} Days ago.
-                  </p>
+                  <p class="card-text">{{ plant.alert.lastWatered }}</p>
                   <p class="card-text">
                     Recommended: Every
                     {{ plant.care.watering.frequency * 7 }} Days
+                  </p>
+                  <p class="card-text">
+                    {{ plant.alert.daysUntil }}
                   </p>
                   <div></div>
                 </li>
