+++ src/app/forest/forest.component.scss
@@ -73,11 +73,7 @@ article{
   background-size: cover;
 }
 
-.card-img-top{
-  height:225px;
-  width:100%;
-  object-fit:cover;
-}
+
 .custom-buttons{
   position:absolute;
   bottom:0;
@@ -91,6 +87,12 @@ article{
 
 .green{
   background-color:lightblue;
+}
+
+.card-img-top {
+  width: 100%;
+  // height: 15vw;
+  // object-fit: cover;
 }
