+++ src/app/plant-new/plant-new.component.html
@@ -1,5 +1,5 @@
 <form [formGroup]="newPlantForm">
-  <div class="content">
+  <div class="container">
     <div class="form-group row col-12">
       <label class="col-sm-4 col-form-label" for="plantName">Plant Name:</label>
       <div class="col-sm-8">
